reduce middleware bundle size below 1MB edge limit

+++ auth.ts
@@ -1,27 +1,17 @@
 import { prisma } from "@/db/prisma";
-
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import credentialProvider from "next-auth/providers/credentials";
 import { compareSync } from "bcrypt-edge";
-import type { NextAuthConfig } from "next-auth";
 import NextAuth from "next-auth";
 import GoogleProvider from "next-auth/providers/google";
 import FacebookProvider from "next-auth/providers/facebook";
 import TwitterProvider from "next-auth/providers/twitter";
 import AppleProvider from "next-auth/providers/apple";
 import { cookies } from "next/headers";
-import { NextResponse } from "next/server";
+import { authConfig } from "@/auth.config";
 
-export const config = {
-  pages: {
-    signIn: "/sign-in",
-    error: "/sign-in",
-  },
-  session: {
-    strategy: "jwt",
-    maxAge: 30 * 24 * 60 * 60, // 30 days
-  },
-
+export const { handlers, auth, signIn, signOut } = NextAuth({
+  ...authConfig,
   adapter: PrismaAdapter(prisma),
   providers: [
     GoogleProvider({
@@ -45,20 +35,12 @@ export const config = {
         email: { type: "email" },
         password: { type: "password" },
       },
-
       async authorize(credentials) {
-        if (credentials == null) {
-          return null;
-        }
+        if (credentials == null) return null;
 
-        // Find user in Database
         const user = await prisma.user.findFirst({
-          where: {
-            email: credentials.email as string,
-          },
+          where: { email: credentials.email as string },
         });
-
-        // Check if user exists and password is match
 
         if (user && user.password) {
           const isMatch = compareSync(
@@ -74,37 +56,26 @@ export const config = {
             };
           }
         }
-
-        // If user not found or password is not match, return null
         return null;
       },
     }),
   ],
   callbacks: {
+    ...authConfig.callbacks,
     async session({ session, user, trigger, token }: any) {
-      // Set the user ID from the token
       session.user.id = token.sub;
       session.user.role = token.role;
       session.user.name = token.name;
-
-      // If there is an update, set the user name
       if (trigger === "update" && token.name) {
         session.user.name = user.name;
       }
       return session;
     },
-
     async jwt({ token, user, trigger, session }: any) {
-      // Assign user fields to token
       if (user) {
         token.role = user.role;
-        token.role = user.role;
-
-        // If user has no name then use the email
         if (user.name === "NO_NAME") {
           token.name = user.email!.split("@")[0];
-
-          // Update database to reflect the token name
           await prisma.user.update({
             where: { id: user.id },
             data: { name: token.name },
@@ -113,18 +84,12 @@ export const config = {
         if (trigger === "signIn" || trigger === "signUp") {
           const cookiesObject = await cookies();
           const sessionCartId = cookiesObject.get("sessionCartId")?.value;
-
           if (sessionCartId) {
             const sessionCart = await prisma.cart.findFirst({
               where: { sessionCartId },
             });
-
             if (sessionCart) {
-              // Delete any existing cart for the user
-              await prisma.cart.deleteMany({
-                where: { userId: user.id },
-              });
-              // Assign new cart
+              await prisma.cart.deleteMany({ where: { userId: user.id } });
               await prisma.cart.update({
                 where: { id: sessionCart.id },
                 data: { userId: user.id },
@@ -133,59 +98,10 @@ export const config = {
           }
         }
       }
-
-      // Handle session Updates
       if (session?.user.name && trigger === "update") {
         token.name = session.user.name;
       }
-
       return token;
     },
-    authorized({ request, auth }: any) {
-      // Array of regex patterns of paths that should be protected
-      const protectedPaths = [
-        /^\/shipping-address(\/.*)?$/,
-        /^\/payment-method(\/.*)?$/,
-        /^\/place-order(\/.*)?$/,
-        /^\/profile(\/.*)?$/,
-        /^\/user(\/.*)?$/,
-        /^\/order(\/.*)?$/,
-        /^\/admin(\/.*)?$/,
-      ];
-
-      // Get pathname from req URL Object
-      const { pathname } = new URL(request.url);
-
-      // Check if use is not authenticated and is trying to access protected path
-      if (!auth && protectedPaths.some((pattern) => pattern.test(pathname))) {
-        // Redirect to sign-in page
-        return NextResponse.redirect(new URL("/sign-in", request.url));
-      }
-
-      // Check for session cart cookie
-      if (!request.cookies.get("sessionCartId")) {
-        // Generate new session cart id cookie
-        const sessionCartId = crypto.randomUUID();
-        //console.log(sessionCartId);
-        // Clone the req headers
-        const newRequestHeaders = new Headers(request.headers);
-
-        // Create new response and add the new headers
-        const response = NextResponse.next({
-          request: {
-            headers: newRequestHeaders,
-          },
-        });
-
-        // Set newly generated sessionCartId in the response cookies
-        response.cookies.set("sessionCartId", sessionCartId);
-
-        return response;
-      } else {
-        return true;
-      }
-    },
   },
-} satisfies NextAuthConfig;
-
-export const { handlers, auth, signIn, signOut } = NextAuth(config);
+});
